Create base class for SSD168x-based devices
