+++ devices/ePaper/Drivers/Ssd168x/RefreshMode.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2022 The nanoFramework project contributors
 // See LICENSE file in the project root for full license information.
 
-namespace Iot.Device.EPaper.Drivers.Ssd1681
+namespace Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681
 {
     /// <summary>
     /// SSD1861 Supported Refresh Modes.
