+++ devices/ePaper/ePaper.nfproj
@@ -37,12 +37,13 @@
     <Compile Include="Buffers\FrameBufferBase.cs" />
     <Compile Include="Buffers\IFrameBuffer.cs" />
     <Compile Include="Drivers\IePaperDisplay.cs" />
-    <Compile Include="Drivers\Ssd1680\Ssd1680.cs" />
-    <Compile Include="Drivers\Ssd1681\Command.cs" />
-    <Compile Include="Drivers\Ssd1681\Ram.cs" />
-    <Compile Include="Drivers\Ssd1681\RefreshMode.cs" />
-    <Compile Include="Drivers\Ssd1681\SleepMode.cs" />
-    <Compile Include="Drivers\Ssd1681\Ssd1681.cs" />
+    <Compile Include="Drivers\Ssd168x\Ssd1680\Ssd1680.cs" />
+    <Compile Include="Drivers\Ssd168x\Command.cs" />
+    <Compile Include="Drivers\Ssd168x\Ram.cs" />
+    <Compile Include="Drivers\Ssd168x\RefreshMode.cs" />
+    <Compile Include="Drivers\Ssd168x\SleepMode.cs" />
+    <Compile Include="Drivers\Ssd168x\Ssd1681\Ssd1681.cs" />
+    <Compile Include="Drivers\Ssd168x\Ssd168x.cs" />
     <Compile Include="Enums\ColorFormat.cs" />
     <Compile Include="Enums\PowerState.cs" />
     <Compile Include="Enums\Rotation.cs" />
